Massively improve coordination between bigbang Py script and bbclient Java.

+++ bbclient/src/com/bbclient/IsEmptyRunnable.java
@@ -3,7 +3,7 @@ package com.bbclient;
 import com.hazelcast.map.IMap;
 
 class IsEmptyRunnable extends IMapMethodRunnable {
-    IsEmptyRunnable(IMap<Long, String> map) {
+    IsEmptyRunnable(IMap<Integer, String> map) {
         super(map,"isEmpty", false);
     }
